pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ocr"
version = "0.1.0"
description = "FastAPI based OCR service with EasyOCR and LLM support"
requires-python = ">=3.10"
dependencies = [
    "fastapi==0.104.1",
    "uvicorn[standard]==0.24.0",
    "python-multipart==0.0.6",
    "easyocr==1.7.1",
    "openai==1.3.5",
    "pillow==10.0.1",
    "pydantic==2.5.0",
    "pydantic-settings==2.1.0",
    "python-dotenv==1.0.0",
    "httpx==0.25.1",
    "pytesseract==0.3.10",
    "pdf2image==1.16.3",
    "PyPDF2==3.0.1",
    "google-generativeai==0.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest==7.4.3",
    "pytest-asyncio==0.21.1",
    "black==23.11.0",
    "isort==5.12.0",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]